RELEASE: bot v0.1.0

+++ db/func/base.py
@@ -2,14 +2,8 @@ from db.core.engine import engine
 from db.core.base import Base
 
 
-async def create_tables():
+async def create_tables(if_not_exist: bool = False):
     async with engine.connect() as conn:
-        await conn.run_sync(Base.metadata.drop_all)
-        await conn.run_sync(Base.metadata.create_all)
-        await conn.commit()
-
-
-async def create_tables_if_not_exists():
-    async with engine.connect() as conn:
+        if not if_not_exist: await conn.run_sync(Base.metadata.drop_all)
         await conn.run_sync(Base.metadata.create_all)
         await conn.commit()
